Conflicts:
	Source/WTF/wtf/Platform.h
	Source/WTF/wtf/text/WTFString.h
	Source/WebCore/bindings/js/JSInspectorFrontendHostCustom.cpp
	Source/WebCore/platform/ContextMenu.h
	Source/WebCore/platform/ContextMenuItem.h
	Source/WebCore/platform/Cursor.h
	Source/WebCore/platform/PlatformKeyboardEvent.h
	Source/WebCore/platform/PlatformMenuDescription.h
	Source/WebCore/platform/PlatformMouseEvent.h
	Source/WebCore/platform/PlatformWheelEvent.h
	Source/WebCore/platform/Widget.h
	Source/WebCore/platform/graphics/BitmapImage.h
	Source/WebCore/platform/graphics/Color.h
	Source/WebCore/platform/graphics/FloatPoint.h
	Source/WebCore/platform/graphics/FloatRect.h
	Source/WebCore/platform/graphics/FontPlatformData.h
	Source/WebCore/platform/graphics/Gradient.h
	Source/WebCore/platform/graphics/GraphicsContext.cpp
	Source/WebCore/platform/graphics/GraphicsContext.h
	Source/WebCore/platform/graphics/IntSize.h
	Source/WebCore/platform/graphics/Path.h
	Source/WebCore/platform/graphics/Pattern.h
	Tools/Scripts/webkitdirs.pm

+++ Source/WebCore/platform/graphics/blackberry/LayerWebKitThread.cpp
@@ -341,7 +341,7 @@ void LayerWebKitThread::removeSublayerOrOverlay(LayerWebKitThread* sublayer)
 
 void LayerWebKitThread::remove(Vector<RefPtr<LayerWebKitThread> >& vector, LayerWebKitThread* sublayer)
 {
-    int foundIndex = vector.find(sublayer);
+    size_t foundIndex = vector.find(sublayer);
     if (foundIndex == notFound)
         return;
 
@@ -359,7 +359,7 @@ void LayerWebKitThread::replaceSublayer(LayerWebKitThread* reference, PassRefPtr
     if (reference == newLayer)
         return;
 
-    int referenceIndex = m_sublayers.find(reference);
+    size_t referenceIndex = m_sublayers.find(reference);
     if (referenceIndex == notFound) {
         ASSERT_NOT_REACHED();
         return;
@@ -393,7 +393,7 @@ void LayerWebKitThread::setBounds(const IntSize& size)
 void LayerWebKitThread::setFrame(const FloatRect& rect)
 {
     if (rect == m_frame)
-      return;
+        return;
 
     m_frame = rect;
     setNeedsDisplay();
